Cleanup test providers
